Update License
Since we want this project to be a source of revenue for the selected students from Mozambique/Burundi, I thought it better to add a more restrictive license. It's still fully open, but this means that if someone picks the hard work of these students they have to keep it open-source.
[Check here for more details on this specific license][0]

[0]:https://choosealicense.com/licenses/gpl-3.0/