Install GUT for unit and integration testing
Trying to have some basic tests so the rest of the project can go smoothly (and we can have a testing framework as we add new features). [Using GUT for testing right now, working in Godot and the IDE][0]. Next step should be for it to run automatically with Github Actions before allowing a merge into master.

[0]: https://github.com/bitwes/Gut/wiki