Update Game.gd
here firstly I made my panel, heartsempty and heartsfull visible on click  of key command to start the game.

on Lose() function I just added the code for verify if lives is equals to zero in other to set it again to one that is the our minimal condition. So if lives is over zero we keep our player playing the same level.

When the player win, I mean: on win() function I just call the method to increase one life on the player

at least I created the ModifyLives(int) function that increment or decrement lives according to the player situation call it whenever I need to increase or decrease life on the player and it works as hoped.