Change HUD from hearts to a player icon
I didn't love the heart, so changed it to a player icon.
Also fixed a bug that was preventing the player from advancing in levels. WE have to create unit tests for the Win() and Lose() function.

Change HUD from hearts to a player icon

I didn't love the heart, so changed it to a player icon.
Also fixed a bug that was preventing the player from advancing in levels. WE have to create unit tests for the Win() and Lose() function.