Updated Tutorial
This commit is multifold, but it basically cleaned up a lot of the animations/graphics and then logic of the tutorial. This now correctly displays the tutorial over the first 5 stages, with support for keyboard and controller. Some logic had to be added to Tutorials.gd to enable/disable and rotate the possible controls (including some code to NOT show the controls on stage 5).

Not the prettiest code but it works.