Fix wall bug
In the conversion to Godot 4 we broke the functionality of rotating the goober when he reaches a wall.

This fixes it by using the Godot functionality of is_on_wall() and depending on how the sprite is turned, adding the inverse default velocity (since it's zero once it reaches the wall).

Also changed the code for checking if the floor is present for simplicity.

The change in the TestIntegrationScene is there so I could test the behavior in scene instead of having to load the entire game (should probably make unit tests for Goober).